requirements.txt
Django==1.9
MySQL-python==1.2.5
Pillow==3.3.0
django-imagekit==3.3
django-modeldict==1.4.1
django-meta==1.3.1